requirements.txt
python-dotenv~=1.0.1
Telethon~=1.36.0
colorama~=0.4.6
httpx~=0.27.0
aiosqlite~=0.20.0
SQLAlchemy~=2.0.31
alembic~=1.13.2
psycopg2~=2.9.9
prefect~=2.19.7